snakegame/requirements.txt
aggdraw==1.3.18.post0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0.post1
Faker==25.8.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
geojson==2.5.0
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-experimental==0.0.62
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.93
lazy_loader==0.4
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
ollama==0.3.0
openai==1.35.3
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psd-tools==1.9.34
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyowm==3.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.6.0
yarl==1.9.4
